Rework the unit tests for sorted table iteration
I consider it dangerous to rely on directly fiddling with the
'internal' sortedNext() iterator function. The tests present
in test/test_luaunit.lua work equally well when using a cleaner
approach that takes sortedPairs() and mimics the behavior of
generic for loops.

This is a preparatory step towards replacing the sortedNext()
iterator with a slighty different implementation. We shouldn't
try to outsmart "for"'s ability to properly manage the state...