Improve on sortedNext() and _table_tostring()
When locating an "out-of-sequence" key value, sortedNext() can now
do better than a naive first matching key search. We know that the
key table is sorted in crossTypeSort() order, so it's possible to
use bisection to locate keys quickly.

The newly gained ability of sortedPairs() to deal nicely with
heterogeneous keys also allows us to avoid the "split" ipairs()
plus pairs() processing logic in _table_tostring().