Support a date format override
The default date format for startDate, set in LuaUnit:startSuite()
and subsequently used by several outputters, depends on the current
system locale. This may be undesirable when trying to create
uniform output, e.g. for the test reference files.

This commit is a one-line change that supports overriding the date
format via setting an environment variable (LUAUNIT_DATEFMT).