Refine test count management
This commit introduces a "passed" test count to display results more
accurately in case of premature abortion (where we can no longer rely
on passedCount = testCount - notPassedCount).

It's also potentially inconsistent to manage separate counters where
we already track all relevant nodes in a Lua table: the corresponding
test count is equal to the number of elements in the table.