Add new LuaUnit option (--error, -e) to quit on errors
Users may wish to abort lengthy test suites in case an actual
error occurs (i.e. when encountering a runtime problem instead
of 'expected' test failures). To keep compatibility with
previous LuaUnit versions, this should be an "opt-in" feature
that has to be requested deliberately.

This commit adds a new command line parameter to achieve this.
With the "-e" ("--error") option set, LuaUnit will (cleanly) stop
test execution in case of any error, and quit with exit code -2.

(A unit test was added that demonstrates and checks this.)