Allow zero margin again for almostEquals()
Also see the previous discussion at
https://github.com/bluebird75/luaunit/issues/33

I've reconsidered this and come to the conclusion that there
might be valid reasons for using a zero margin parameter with
assertAlmostEquals() or assertNotAlmostEquals(). E.g. if the
user has to deal with rounding errors within our "boost" interval
(EPSILON), but doesn't want (to add) any additional "margin",
then passing 0 actually makes sense.

assertEquals() and assertNotEquals() do not provide an adequate
replacement in that particular case.