Rework the matching logic in _is_table_equals()
Currently the function is checking many values (actual[k] vs.
expected[k]) *twice* - for those keys that are present in both
tables.

This is disadvantageous, especially when we try to keep track of
possible recursion loops; since the double-check would cause two
recursive invocations of _is_table_equals() with swapped, but
otherwise identical arguments. This leads to "false positives"
when testing for tables that were already visited.

Fortunately there's a simple way to avoid the second recursion
completely, if we just apply a similar logic to what we're already
doing with table-type keys: Keep track of the keys that were
matched from/in "actual", and remove them successively when
iterating "expected".

With this change applied, all unit tests still pass unmodified.