Fix XML schema examples
The example files in the junitxml subdirectory had some problems.
example-bamboo-1.xml contained a series of dots (possibly indicating
an omission/removal of further <property> lines?) that would prevent
proper validation. And example-bamboo-2.xml seemed to have lost a
'<' character at the beginning, making it invalid XML.

This patch fixes that and adds a simple Makefile to the directory,
so that xmllint checks can be easily run if someone's interested.