Improve _is_table_equals function for some recursive tables
Continue descending when the pair (actual,expected) has not yet been visited.
That is, both actual and expected could have been visited when paired with
other tables, just not this specific pair.
Since actual and expected tables are finite, the number of subtables they
consist of is finite as well, thus the number of combinations of
(actual,expected) is finite as well, and the recursion will stop eventually.

Also moved some checks earlier, to avoid unnecessary recursion in special cases.

Changes the output of some test cases (that were previously considering
asymetrically looping tables to be not equal).