Have a "smart" table to keep better track of recursions
This patch picks up an excellent proposal of AlfonZ42 to make use
of the actual (actual,expected) pairings when trying to keep track
of the recursions in _is_table_equals().
See https://github.com/bluebird75/luaunit/pull/57#issuecomment-225295770

However, I would like to use a slightly different approach - I've
chosen to extend the "recursions" table with two methods that help
with this task: recursions:cached() can be used to query the state
of an (actual,expected) pair, and recursions:store() provides a
straightforward way to set this state.