Dynamically calculate a default EPSILON, depending on precision
Our former EPSILON is meant to compensate for (Lua-internal)
inaccuracies, like rounding errors, when doing `almostEquals()`
tests.

This means that its value gets affected by the numerical precision
that the underlying Lua implementation offers: EPSILON needs to be
greater if only "single" precision is available, something that
we now explicitly arrange for in our CI tests.

I have therefore added a 'dynamic' selection of EPSILON. The default
value represents twice the [machine epsilon] for double precision
(which is common in Lua). In case this turns out to be insufficient,
twice the machine epsilon for "single" precision will be used instead.

[machine epsilon]: https://en.wikipedia.org/wiki/Machine_epsilon