Remove some unreachable code lines from _is_table_equals()
Coverage analysis indicates that the test for an "empty"
actualKeysMatched table always succeeds, and the enclosed
return statement never gets executed.

This is to be expected, as for any given (non table-type)
key `k` that is present in `actual` but not in `expected`,
our earlier value-based test would already fail. We compare
actual[k] against (expected[k] == nil), which won't pass.

Fixed this by removing the unnecessary code.