Flush output in the default verbosity
In the default verbosity, LuaUnit outputs one character such as "."
and "E". We can't know the one character progress until all tests are
finished or the stdout internal buffer is fulled because the one
characters are buffered in the stdout internal buffer.

We can the one character progress immediately by flushing the stdout
internal buffer.