Improve table compare cycle detection and remove option TABLE_EQUALS_KEYBYCONTENT
closes #116

The table comparison now correctly deals with table containing cycles of different structure. Previously, it would return true on table with cycles, even if the cycles were of different style (like 2 table cycles vs 3 table cycles).

The option TABLE_EQUALS_KEYBYCONTENT is removed because it does not make sense : lua treats table as keys by reference.

An example that does not make sense : t1={ [{1}]=true, [{1}]=false }
Two keys have exactly the same content, but a different value, within the same table. We would not know which one to pick for comparison with another table so the result of comparison would be non deterministic.