options: allow specifying the output format via the environment
Specifically, treat LUAUNIT_DEFAULT_OUTPUT in the OS environment as if it
were passed via -o, unless an -o option is actually specified -- in that
case, the argument always overrides.

This is sometimes necessary when trying to seamlessly adapt luaunit tests
to some test drivers.  e.g., FreeBSD uses Kyua to drive its tests, which
generally wants to just execute the program/script that it's pointed at and
observe the results.  Kyua knows about the TAP format but not the plain
format used by luaunit, so this allows a simple adaptation:

 #!/usr/bin/env -S LUAUNIT_DEFAULT_OUTPUT=TAP /usr/local/bin/lua53

In this case, the shebang line is added automatically by the build system
that's generating the final test script, so the source test script doesn't
need to know that this is happening -- it just uses luaunit as it normally
would, and one can either execute it manually to get the stock plain format
or through the framework, which goes through an incantation like the above
to hint the format to TAP.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>